requirements.txt
pandas_gbq
google.cloud
pytest
seaborn
numpy==1.23.5
pandas==1.5.3
scipy==1.10.0
scikit-learn==1.3.1
#google-cloud-bigquery
google-cloud-storage
python-io
python-dotenv
